# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IRUSA Smart Event & Donor Intelligence Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /donors:
    get:
      operationId: listDonors
      tags: [donors]
      summary: List all donors
      parameters:
        - in: query
          name: donorCategory
          schema:
            type: string
        - in: query
          name: personalityType
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Donors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donor"
    post:
      operationId: createDonor
      tags: [donors]
      summary: Create donor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonorBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
  /donors/{id}:
    get:
      operationId: getDonor
      tags: [donors]
      summary: Get donor by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "404":
          description: Not found
    patch:
      operationId: updateDonor
      tags: [donors]
      summary: Update donor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDonorBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
  /donors/{id}/profile:
    get:
      operationId: getDonorProfile
      tags: [donors]
      summary: Get full donor profile with insights, donations, and events attended
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full donor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonorProfile"
  /donors/{id}/recommendations:
    get:
      operationId: getDonorRecommendations
      tags: [donors]
      summary: Get AI-driven next-step recommendations for a donor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  # ==================== DONATIONS ====================
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
        location:
          type: string
        masjidPartner:
          type: ["string", "null"]
        eventType:
          type: string
        campaign:
          type: ["string", "null"]
        organizer:
          type: ["string", "null"]
        status:
          type: string
        estimatedAttendees:
          type: ["integer", "null"]
        actualAttendees:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - date
        - location
        - eventType
        - status
        - createdAt
        - updatedAt
...
    Donor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        totalDonated:
          type: number
        averageDonation:
          type: number
        donationCount:
          type: integer
        firstDonationDate:
          type: ["string", "null"]
        lastDonationDate:
          type: ["string", "null"]
        donorCategory:
          type: string
        donorPersonalityType:
          type: ["string", "null"]
        preferredContactFrequency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - totalDonated
        - averageDonation
        - donationCount
        - donorCategory
        - createdAt
        - updatedAt
    CreateDonorBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        donorCategory:
          type: string
          default: "one-time"
        donorPersonalityType:
          type: ["string", "null"]
        preferredContactFrequency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
    UpdateDonorBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        donorCategory:
          type: string
        donorPersonalityType:
          type: ["string", "null"]
        preferredContactFrequency:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    DonorProfile:
      type: object
      properties:
        donor:
          $ref: "#/components/schemas/Donor"
        donations:
          type: array
          items:
            $ref: "#/components/schemas/Donation"
        eventsAttended:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        topCauses:
          type: array
          items:
            type: string
        isRamadanGiver:
          type: boolean
        isEmergencyResponder:
          type: boolean
        givingFrequencyScore:
          type: number
        recencyScore:
          type: number
        engagementRiskScore:
          type: number
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
      required:
        - donor
        - donations
        - eventsAttended
        - topCauses
        - isRamadanGiver
        - isEmergencyResponder
        - givingFrequencyScore
        - recencyScore
        - engagementRiskScore
        - recommendations
    Recommendation:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        reason:
          type: string
        urgency:
          type: string
        tone:
          type: string
        suggestedSubject:
          type: ["string", "null"]
        suggestedMessage:
          type: ["string", "null"]
      required:
        - id
        - action
        - reason
        - urgency
        - tone

    # ==================== DONATION ====================
    Donation:
      type: object
      properties:
        id:
          type: integer
        donorId:
          type: integer
        eventId:
          type: ["integer", "null"]
        date:
          type: string
        amount:
          type: number
        cause:
          type: ["string", "null"]
        campaign:
          type: ["string", "null"]
        season:
          type: ["string", "null"]
        donationType:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - donorId
        - date
        - amount
        - donationType
        - createdAt
...