# lib/db/src/schema/donors.ts
import { pgTable, text, serial, timestamp, integer, numeric } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const donorsTable = pgTable("donors", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  email: text("email"),
  phone: text("phone"),
  location: text("location"),
  totalDonated: numeric("total_donated", { precision: 12, scale: 2 }).notNull().default("0"),
  averageDonation: numeric("average_donation", { precision: 12, scale: 2 }).notNull().default("0"),
  donationCount: integer("donation_count").notNull().default(0),
  firstDonationDate: text("first_donation_date"),
  lastDonationDate: text("last_donation_date"),
  donorCategory: text("donor_category").notNull().default("one-time"),
  donorPersonalityType: text("donor_personality_type"),
  preferredContactFrequency: text("preferred_contact_frequency"),
  notes: text("notes"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertDonorSchema = createInsertSchema(donorsTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertDonor = z.infer<typeof insertDonorSchema>;
export type Donor = typeof donorsTable.$inferSelect;